bench/ipa: unify benchmarks
