+++ benches/benchmarks/mod.rs
@@ -1,4 +1,3 @@
 pub mod ipa_verify;
-pub mod ipa_verify_multi_exp;
 pub mod multipoint_prove;
 pub mod multipoint_verify;
